requirements.txt
# GUI Framework
customtkinter>=5.2.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# AI & Background Removal
rembg>=2.0.50
onnxruntime>=1.15.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
huggingface-hub>=0.16.0

# Engineering / CAD / CNC
ezdxf>=1.1.0

# System Utilities
packaging>=23.0
filetype>=1.2.0